+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * [CONTRIBUTING] Document our pull request expectations in CONTRIBUTING.md. [Jeff Verkoeyen](https://github.com/jverkoey)
 * [Documentation] Added Swift example for Typography. [Peter Friese](https://github.com/peterfriese)
 * [Switch] Removed internal docs that were pretending to be public docs. [Jeff Verkoeyen](https://github.com/jverkoey)
+* [Samples] Added Swift sample for buttons. [Peter Friese](https://github.com/peterfriese)
 
 ##### Bug Fixes